Fetch identifier value from pan-cancer dataset

Identifier includes gene/probe etc.

get_ccle_cn_value(identifier)

get_ccle_gene_value(identifier)

get_ccle_protein_value(identifier)

get_ccle_mutation_status(identifier)

get_pancan_value(
  identifier,
  subtype = NULL,
  dataset = NULL,
  host = available_hosts(),
  samples = NULL
)

get_pancan_gene_value(identifier)

get_pancan_transcript_value(identifier)

get_pancan_protein_value(identifier)

get_pancan_mutation_status(identifier)

get_pancan_cn_value(identifier)

get_pancan_methylation_value(identifier, type = c("450K", "27K"))

get_pancan_miRNA_value(identifier)

Arguments

identifier
a length-1 character representing a gene symbol, ensembl gene id, or probe id. Gene symbol is highly recommended.

subtype
a length-1 chracter representing a regular expression for matching DataSubtype column of UCSCXenaTools::XenaData.

dataset
a length-1 chracter representing a regular expression for matching XenaDatasets of UCSCXenaTools::XenaData.

host
a character vector representing host name(s), e.g. "toilHub".

samples
a character vector representing samples want to be returned.

type
methylation type, one of "450K" and "27K".

Value

a named vector or list

Functions

get_ccle_cn_value: Fetch copy number value from CCLE dataset

get_ccle_gene_value: Fetch gene expression value from CCLE dataset

get_ccle_protein_value: Fetch gene protein expression value from CCLE dataset

get_ccle_mutation_status: Fetch gene mutation info from CCLE dataset

get_pancan_value: Fetch identifier value from pan-cancer dataset

get_pancan_gene_value: Fetch gene expression value from pan-cancer dataset

get_pancan_transcript_value: Fetch gene transcript expression value from pan-cancer dataset

get_pancan_protein_value: Fetch protein expression value from pan-cancer dataset

get_pancan_mutation_status: Fetch mutation status value from pan-cancer dataset

get_pancan_cn_value: Fetch gene copy number value from pan-cancer dataset processed by GISTIC 2.0

get_pancan_methylation_value: Fetch gene expression value from CCLE dataset

get_pancan_miRNA_value: Fetch miRNA expression value from pan-cancer dataset

Examples

# \donttest{
# Fetch TP53 expression value from pan-cancer dataset
t1 <- get_pancan_value("TP53",
  dataset = "TcgaTargetGtex_rsem_isoform_tpm",
  host = "toilHub"
)
t2 <- get_pancan_gene_value("TP53")
t3 <- get_pancan_protein_value("AKT")
t4 <- get_pancan_mutation_status("TP53")
t5 <- get_pancan_cn_value("TP53")
# }