Visualize single gene expression from toil data hub

Visualize single gene expression from toil data hub

vis_toil_gene(
  data,
  x = "primary_site",
  y = "expression",
  color = "sample_type",
  palette = "jco",
  xlab = "Primary site",
  ylab = "Expression",
  title = NULL,
  facet.by = NULL,
  angle.x = 45,
  ...
)

Arguments

data
a data frame

x
character string containing the name of x variable.

y
character vector containing one or more variables to plot

color
outline color.

palette
the color palette to be used for coloring or filling by groups. Allowed values include "grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...; or custom color palette e.g. c("blue", "red"); and scientific journal palettes from ggsci R package, e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".

xlab
character vector specifying x axis labels. Use xlab = FALSE to hide xlab.

ylab
character vector specifying y axis labels. Use ylab = FALSE to hide ylab.

title
plot main title.

facet.by
character vector, of length 1 or 2, specifying grouping variables for faceting the plot into multiple panels. Should be in the data.

angle.x
angle for x lab

...
other parameters passing to ggpubr::ggboxplot()

Value

a ggplot object