Visualize Single Gene Univariable Cox Result from Toil Data Hub

Visualize Single Gene Univariable Cox Result from Toil Data Hub

vis_unicox_tree(
  Gene = "TP53",
  measure = "OS",
  data_type = "mRNA",
  threshold = 0.5,
  values = c("grey", "#E31A1C", "#377DB8")
)

Arguments

Gene

a molecular identifier (e.g., "TP53") or a formula specifying genomic signature ("TP53 + 2 * KRAS - 1.3 * PTEN").

measure

a survival measure, e.g. "OS".

data_type

choose gene profile type, including "mRNA","transcript","methylation","miRNA","protein","cnv_gistic2"

threshold

a expression cutoff, 0.5 for median.

values

the color to fill tumor or normal

Value

a ggplot object

Examples

if (FALSE) {
p <- vis_unicox_tree(Gene = "TP53")
}